Added file level uploading
Created  a level file as well as modified the readME to include how to create shapes for the level. I tested it a bit but there might still be some edge case issues. I also added code to draw in the shapes that were read from the files so we could make sure it was looking correct.